+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Generate the JSON Schema for the `website-stalker.yaml` with `website-stalker json-schema`.
+- new editor: `css_tag_replace`
 - Support zstd response body decompression
 
 ## [0.25.1] - 2025-01-30